Rename routes and remove unused deps
