+++ src/server/app.js
@@ -2,10 +2,9 @@ const express = require('express')
 const path = require('path')
 const devMiddleware = require('webpack-dev-middleware')
 const hotMiddleware = require('webpack-hot-middleware')
-const expressHandlebars = require('express-handlebars')
 const webpack = require('webpack')
 const webpackConfig = require('../../webpack.config.js')
-const videoRoute = require('./routes/video')
+const videoRoute = require('./routes/embed')
 
 const compiler = webpack(webpackConfig)
 const app = express()
@@ -18,10 +17,7 @@ app.use(
 app.use(hotMiddleware(compiler))
 
 app.use(express.static(path.resolve(__dirname, '../../', 'build')))
-app.engine('handlebars', expressHandlebars())
-app.set('view engine', 'handlebars')
-app.set('views', path.join(__dirname, 'views'))
-app.get('/video/:id', videoRoute)
+app.get('/embed/:id', videoRoute)
 app.get('*', (req, res) => {
   res.sendFile(path.resolve(__dirname, '../../', 'build', 'index.html'))
 })
