--- src/server/views/video.handlebars
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>{{title}}</title>
-  <link rel="stylesheet" href="/embed.css">
-  <script src="/bundle.js"></script>
-  <meta name="og:title" content="{{title}}"/>
-  <meta name="og:image" content="{{thumbnailUrl}}"/>
-  <meta name="og:description" content="{{description}}"/>
-  <meta name="og:url" content="{{url}}"/>
-</head>
-<body>
-  <div id="player"></div>
-  <script type="application/javascript">
-    window.paratii.mediaPlayer({
-      selector: '#player',
-      source: 'https://gateway.paratii.video/ipfs/{{id}}/master.m3u8',
-      mimeType: 'video/mp4',
-      ipfsHash: '{{id}}'
-    })
-  </script>
-</body>
-</html>
